+++ ndc-lambda-sdk/src/sdk.ts
@@ -1,3 +1,3 @@
-import { JSONValue } from "./schema";
-
-export { JSONValue };
+export { JSONValue } from "./schema";
+export { ConnectorError, BadRequest, Forbidden, Conflict, UnprocessableContent, InternalServerError, NotSupported, BadGateway } from "@hasura/ndc-sdk-typescript";
+export { ErrorDetails, getErrorDetails } from "./execution";