+++ ndc-lambda-sdk/test/init.ts
@@ -1,0 +1,8 @@
+// This file is hooked into mocha via .mocharc.json
+
+import chai from "chai";
+import chaiAsPromised from "chai-as-promised"
+
+export const mochaGlobalSetup = function() {
+  chai.use(chaiAsPromised);
+};